Update to use visual studio's C99 stdbool library
According to documentation, VS2013 (dev12) had support for the c99 stdbool
library. The codebase (built with dev6) is assuming visual studio doesn't
have support. Which was true at the time.

I have access to VS2017 but not VS2013 at the moment so I've tested the
build against that and I'm guarding against that.

This would probably be worth trying to push upstream with a version check
of >= 1800 to do it "right".