[BROKEN] Use standard snprintf and vsnprintf
The code base used custom written versions of snprintf and vsnprintf. As of
the UCRT that runs against Visual Studio 2015 and Windows 10, both of
those are implemented to be C99 standards compliant.

Detecting if we're using a recent VC++ compiler and if so, using the UCRT
functions instead.

This is probably worth attempting to push upstream with appropriate checks
(against VS2015 aka < version 1900)