[FIXED] WSPiAPIWrap.h workaround for MinGW breaks
The workaround for MinGW for WSPIAPI on Windows appears to break in the
linker now whereas according to comments it used to be fine. Since this
was for MinGW explicitly from what I saw in the code, I bypassed the
function redefintions.

This fixes the build and there should now be a build of uqmdebug.exe from
within VS2017.