[BROKEN] Added a UWP project.
This UWP project currently depends on a SDL project that is not part of
the same source root. Need to decide what (if anything) to do about that.
At some point the dream is that the SDL 2 project is a WinRT Component
inside a VCPKG lib but I still need to handle it in the meantime.