[BROKEN] Update to avoid the UNICODE predefine
As far as I can tell, when building UWP you get two predefines that you
cannot undo. _UNICODE and UNICODE. Therefore, when a typedef is named
UNICODE, everything is broken. This fixes that in a brutish way (see rest
of comment)

There is a big TODO here. I want to debug and understand the UNICODE
handling of this code-base. I'm assuming the strategy is UTF-8 in the
codebase and either UTF8 or UTF16 at the edges but it isn't clear just
yet. Once that is sorted, I'll add to the docs for future users and at
that time likely remove any remaining dangling references to the UNICODE
type.