[BROKEN] Fix bugs for a stricter compile (Signing)
This code was attempting to negate values it claimed were unsigned. The
compiler pointed out that was probably an error.

I've modified the code to be signed. In addition, the technique wasn't the
most straight forward rendering of a subtraction I've ever seen so I
commented it and renamed a field that was misleading (mask vs. bitarray).