Fixup logging, simplify the file system codepaths
There are a few changes in here.

First is re-enabling logging. I don't have a console window to pipe to so
a file is the right choice. Plus I might be able to grab that latter if
there are failures in the wild. TBD.

Second, is removing some functions near save to simplify the work i need
to do to get the filesystem code updated for appx.

Finally, fixing a bug in the addons part of the options structure that
will cause a crash.