Make the Config directory a R/W dir.
The original authros require that the melee dir (/teams/) be relative to
the config dir. Rather than fix that up and see where that takes me, I
just decided to put the config in the a dir that is R/W instead of RO. So
I flipped this around. Long story short, Super Melee works now.