Refactoring and removing global variable: SetupTab
SetupTab was a static variable holding onto strings for the setup menu
from the resource system. Technically since it was allocated in the
static segment and I moved it to being stack allocated this might be a
perf regression but I only found 2 uses.

Also, making the code flow a little simpler by moving to a guard pattern
(fewer if/else clauses).

In addition, one of the ifs (the one around clean_up_widgets) seemed
superflous after scanning the code. The thread and process would've
exited before hitting that if check from what I saw reading the code. I
don't have a strong test validating that.